Refactor template wizard into package and add tests
